# instalacao/api/storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/enderecos": {
            "get": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Get a listing of the Enderecos.",
                "description": "Get all Enderecos",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Endereco"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Store a newly created Endereco in storage",
                "description": "Store Endereco",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Endereco that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Endereco"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Endereco"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/enderecos/{id}": {
            "get": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Display the specified Endereco",
                "description": "Get Endereco",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Endereco",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Endereco"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Update the specified Endereco in storage",
                "description": "Update Endereco",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Endereco",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Endereco that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Endereco"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Endereco"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Remove the specified Endereco from storage",
                "description": "Delete Endereco",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Endereco",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/instalacaos": {
            "get": {
                "tags": [
                    "Instalacao"
                ],
                "summary": "Get a listing of the Instalacaos.",
                "description": "Get all Instalacaos",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Instalacao"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Instalacao"
                ],
                "summary": "Store a newly created Instalacao in storage",
                "description": "Store Instalacao",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Instalacao that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Instalacao"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Instalacao"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/instalacaos/{id}": {
            "get": {
                "tags": [
                    "Instalacao"
                ],
                "summary": "Display the specified Instalacao",
                "description": "Get Instalacao",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Instalacao",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Instalacao"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Instalacao"
                ],
                "summary": "Update the specified Instalacao in storage",
                "description": "Update Instalacao",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Instalacao",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Instalacao that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Instalacao"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Instalacao"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Instalacao"
                ],
                "summary": "Remove the specified Instalacao from storage",
                "description": "Delete Instalacao",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Instalacao",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Endereco": {
            "required": [
                "cidade",
                "bairro",
                "estado",
                "cep"
            ],
            "properties": {
                "idEndereco": {
                    "description": "idEndereco",
                    "type": "integer",
                    "format": "int32"
                },
                "cidade": {
                    "description": "cidade",
                    "type": "string"
                },
                "bairro": {
                    "description": "bairro",
                    "type": "string"
                },
                "estado": {
                    "description": "estado",
                    "type": "string"
                },
                "cep": {
                    "description": "cep",
                    "type": "string"
                }
            }
        },
        "Instalacao": {
            "required": [
                "nome",
                "idEndereco"
            ],
            "properties": {
                "idInstalacao": {
                    "description": "idInstalacao",
                    "type": "integer",
                    "format": "int32"
                },
                "nome": {
                    "description": "nome",
                    "type": "string"
                },
                "idEndereco": {
                    "description": "idEndereco",
                    "type": "integer",
                    "format": "int32"
                }
            }
        }
    }
}